fix build postinst/prerm error on firmware builds.

+++ Makefile
@@ -38,6 +38,8 @@ endef
 define Package/$(PKG_NAME)/preinst
 #!/bin/sh
 
+[ -z "$${IPKG_INSTROOT}" ] || exit 0
+
 mkdir -p /usr/share/clashbackup 2>/dev/null
 
 
@@ -96,6 +98,9 @@ endef
 
 define Package/$(PKG_NAME)/postinst
 #!/bin/sh
+
+[ -z "$${IPKG_INSTROOT}" ] || exit 0
+
 rm -rf /tmp/luci*  
 
 if [ -f "/etc/config/clash.bak" ]; then
